feat: avance backend actualizado 28 abril

+++ src/main/java/com/cibertec/yachay/services/OpcionRespuestaService.java
@@ -5,6 +5,7 @@ import com.cibertec.yachay.repositories.OpcionRespuestaRepositorio;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -25,21 +26,78 @@ public class OpcionRespuestaService {
                 .orElseThrow(() -> new RuntimeException("Opción no encontrada con id: " + id));
     }
 
-    public OpcionRespuesta crear(OpcionRespuesta opcion) {
-        return opcionRespuestaRepositorio.save(opcion);
+    public OpcionRespuesta crear(
+            OpcionRespuesta opcion
+    ){
+
+        Long idPregunta=
+                opcion.getPregunta()
+                        .getIdPregunta();
+
+        long cantidad=
+                opcionRespuestaRepositorio
+                        .countByPreguntaIdPregunta(
+                                idPregunta
+                        );
+
+        if(cantidad>=4){
+
+            throw new RuntimeException(
+                    "Solo 4 opciones por pregunta"
+            );
+
+        }
+
+        return opcionRespuestaRepositorio.save(
+                opcion
+        );
+
     }
 
-    public OpcionRespuesta actualizar(Long id, OpcionRespuesta opcion) {
-        OpcionRespuesta existente = obtener(id);
-        existente.setTexto(opcion.getTexto());
-        existente.setPeso(opcion.getPeso());
-        existente.setEsCorrecta(opcion.getEsCorrecta());
-        existente.setPregunta(opcion.getPregunta());
-        existente.setPreguntaSiguiente(opcion.getPreguntaSiguiente());
-        return opcionRespuestaRepositorio.save(existente);
+    public OpcionRespuesta actualizar(
+            Long id,
+            OpcionRespuesta opcion
+    ){
+
+        OpcionRespuesta existente=
+                obtener(id);
+
+        existente.setTexto(
+                opcion.getTexto()
+        );
+
+        existente.setPeso(
+                opcion.getPeso()
+        );
+
+        existente.setEsCorrecta(
+                opcion.getEsCorrecta()
+        );
+
+        existente.setPregunta(
+                opcion.getPregunta()
+        );
+
+        return opcionRespuestaRepositorio.save(
+                existente
+        );
+
     }
 
-    public void eliminar(Long id) {
+
+    @Transactional
+    public void eliminar(Long id){
+
+        if(
+                !opcionRespuestaRepositorio.existsById(id)
+        ){
+            throw new RuntimeException(
+                    "Opción no existe"
+            );
+        }
+
         opcionRespuestaRepositorio.deleteById(id);
+
     }
+
 }
